# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Quiz API",
    "description": "API для системы онлайн-квизов с поддержкой категорий, вопросов, ответов и попыток прохождения",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Программирование"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
      "Quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "JavaScript Основы"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Тест на знание основ JavaScript"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440000"
          },
          "timeLimit": {
            "type": "integer",
            "nullable": true,
            "description": "Время в секундах",
            "example": 1800
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        }
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "880e8400-e29b-41d4-a716-446655440000"
          },
          "quizId": {
            "type": "string",
            "format": "uuid",
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          "text": {
            "type": "string",
            "example": "Что такое замыкание в JavaScript?"
          },
          "order": {
            "type": "integer",
            "example": 1
          },
          "points": {
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          }
        }
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "990e8400-e29b-41d4-a716-446655440000"
          },
          "questionId": {
            "type": "string",
            "format": "uuid",
            "example": "880e8400-e29b-41d4-a716-446655440000"
          },
          "text": {
            "type": "string",
            "example": "Функция, которая имеет доступ к переменным внешней функции"
          },
          "isCorrect": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Получить список всех категорий",
        "responses": {
          "200": {
            "description": "Список категорий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                },
                "example": [
                  {
                    "id": "660e8400-e29b-41d4-a716-446655440000",
                    "name": "Программирование",
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  },
                  {
                    "id": "660e8400-e29b-41d4-a716-446655440001",
                    "name": "Математика",
                    "createdAt": "2024-01-15T11:00:00.000Z",
                    "updatedAt": "2024-01-15T11:00:00.000Z"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Создать новую категорию",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Наука"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Категория успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": "660e8400-e29b-41d4-a716-446655440002",
                  "name": "Наука",
                  "createdAt": "2024-01-15T12:00:00.000Z",
                  "updatedAt": "2024-01-15T12:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Получить категорию по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Категория найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": "660e8400-e29b-41d4-a716-446655440000",
                  "name": "Программирование",
                  "createdAt": "2024-01-15T10:30:00.000Z",
                  "updatedAt": "2024-01-15T10:30:00.000Z"
                }
              }
            }
          },
          "404": {
            "description": "Категория не найдена"
          }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Обновить категорию",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Программирование и разработка"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Категория успешно обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                },
                "example": {
                  "id": "660e8400-e29b-41d4-a716-446655440000",
                  "name": "Программирование и разработка",
                  "createdAt": "2024-01-15T10:30:00.000Z",
                  "updatedAt": "2024-01-15T13:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Категория не найдена"
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Удалить категорию",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Категория успешно удалена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Категория не найдена"
          }
        }
      }
    },
    "/quizzes": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Получить список всех квизов",
        "responses": {
          "200": {
            "description": "Список квизов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                },
                "example": [
                  {
                    "id": "770e8400-e29b-41d4-a716-446655440000",
                    "title": "JavaScript Основы",
                    "description": "Тест на знание основ JavaScript",
                    "categoryId": "660e8400-e29b-41d4-a716-446655440000",
                    "timeLimit": 1800,
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z",
                    "category": {
                      "id": "660e8400-e29b-41d4-a716-446655440000",
                      "name": "Программирование",
                      "createdAt": "2024-01-15T10:30:00.000Z",
                      "updatedAt": "2024-01-15T10:30:00.000Z"
                    }
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Quizzes"],
        "summary": "Создать новый квиз",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "categoryId"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Python для начинающих"
                  },
                  "description": {
                    "type": "string",
                    "example": "Тест на знание основ Python"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                  },
                  "timeLimit": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Время в секундах",
                    "example": 3600
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Квиз успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          }
        }
      }
    },
    "/quizzes/search": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Поиск квизов",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "JavaScript"
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/{id}": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Получить квиз по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Квиз найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                },
                "example": {
                  "id": "770e8400-e29b-41d4-a716-446655440000",
                  "title": "JavaScript Основы",
                  "description": "Тест на знание основ JavaScript",
                  "categoryId": "660e8400-e29b-41d4-a716-446655440000",
                  "timeLimit": 1800,
                  "createdAt": "2024-01-15T10:30:00.000Z",
                  "updatedAt": "2024-01-15T10:30:00.000Z",
                  "category": {
                    "id": "660e8400-e29b-41d4-a716-446655440000",
                    "name": "Программирование",
                    "createdAt": "2024-01-15T10:30:00.000Z",
                    "updatedAt": "2024-01-15T10:30:00.000Z"
                  },
                  "questions": [
                    {
                      "id": "880e8400-e29b-41d4-a716-446655440000",
                      "quizId": "770e8400-e29b-41d4-a716-446655440000",
                      "text": "Что такое замыкание в JavaScript?",
                      "order": 1,
                      "points": 1,
                      "createdAt": "2024-01-15T10:30:00.000Z",
                      "updatedAt": "2024-01-15T10:30:00.000Z",
                      "answers": [
                        {
                          "id": "990e8400-e29b-41d4-a716-446655440000",
                          "questionId": "880e8400-e29b-41d4-a716-446655440000",
                          "text": "Функция, которая имеет доступ к переменным внешней функции",
                          "isCorrect": true,
                          "createdAt": "2024-01-15T10:30:00.000Z",
                          "updatedAt": "2024-01-15T10:30:00.000Z"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Квиз не найден"
          }
        }
      },
      "put": {
        "tags": ["Quizzes"],
        "summary": "Обновить квиз",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "JavaScript Продвинутый"
                  },
                  "description": {
                    "type": "string",
                    "example": "Тест на знание продвинутых концепций JavaScript"
                  },
                  "categoryId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "660e8400-e29b-41d4-a716-446655440000"
                  },
                  "timeLimit": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2400
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Квиз успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Квиз не найден"
          }
        }
      },
      "delete": {
        "tags": ["Quizzes"],
        "summary": "Удалить квиз",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Квиз успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "404": {
            "description": "Квиз не найден"
          }
        }
      }
    },
    "/quizzes/{id}/leaderboard": {
      "get": {
        "tags": ["Quizzes"],
        "summary": "Получить лидерборд квиза",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "770e8400-e29b-41d4-a716-446655440000"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Лидерборд квиза",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "score": {
                        "type": "integer"
                      },
                      "maxScore": {
                        "type": "integer"
                      },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "user": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": "aa0e8400-e29b-41d4-a716-446655440000",
                    "score": 10,
                    "maxScore": 10,
                    "completedAt": "2024-01-15T11:00:00.000Z",
                    "user": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "username": "johndoe",
                      "email": "user@example.com"
                    }
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Квиз не найден"
          }
        }
      }
    },
...
  },
...
}